pyproject.toml
[project]
name = "dfautotrans"
version = "0.1.0"
description = "Dead Frontier Auto Trading System - A Python-based automated trading bot for Dead Frontier marketplace"
readme = "README.md"
authors = [
    { name = "LF\\Tommy.Yeung", email = "tommy.yeung@lungfung.hk" }
]
requires-python = ">=3.11"
dependencies = [
    "playwright>=1.40.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "aiosqlite>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[project.scripts]
dfautotrans = "dfautotrans.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/dfautotrans"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
mypy_path = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
